requirements.txt
torch==2.4.1
numpy==1.26.4
pillow==10.4.0
transformers==4.50.3
requests==2.32.3
tqdm==4.66.5
pandas==2.1.4
selenium==4.23.0
setuptools==68.2.2
open_clip_torch
neptune-client
pytorch-lightning
colpali